refactor VLMs and LLMs to be under same class
