+++ mypy.ini
@@ -97,3 +97,9 @@ ignore_missing_imports = True
 
 [mypy-ImageHash.*]
 ignore_missing_imports = True
+
+[mypy-google.*]
+ignore_missing_imports = True
+
+[mypy-google.generativeai.*]
+ignore_missing_imports = True
